pkg/destroy/gcp: discover cloud controller resources prior to deletion
Previously, there was significant duplication between deletion of
the external load balancer resources and the internal ones. This change
cleans up the duplication and postpones deletion of these resources
until later in the process.
